+++ base/src/main/java/io/github/sds100/keymapper/base/trigger/ConfigTriggerDelegate.kt
@@ -479,10 +479,6 @@ class ConfigTriggerDelegate {
         return trigger.copy(longPressDoubleVibration = enabled).validate()
     }
 
-    fun setTriggerFromOtherAppsEnabled(trigger: Trigger, enabled: Boolean): Trigger {
-        return trigger.copy(triggerFromOtherApps = enabled).validate()
-    }
-
     fun setShowToastEnabled(trigger: Trigger, enabled: Boolean): Trigger {
         return trigger.copy(showToast = enabled).validate()
     }